requirements.txt
appdirs==1.4.3
asgiref==3.3.1
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.6
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-filter==2.4.0
django-serializer==0.0.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.14.0
drf-writable-nested==0.6.2
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
Jinja2==2.11.3
jsonschema==3.2.0
jwt==1.2.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.1.2
progress==1.5
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.17.3
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
swagger-ui-py==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
